+++ src/deploy/functions/runtimes/discovery/v1alpha1.ts
@@ -19,6 +19,7 @@ export type ManifestEndpoint = backend.ServiceConfiguration &
   Partial<backend.CallableTriggered> &
   Partial<backend.EventTriggered> &
   Partial<backend.TaskQueueTriggered> &
+  Partial<backend.BlockingTriggered> &
   Partial<backend.ScheduleTriggered> & {
     region?: string[];
     entryPoint: string;
@@ -84,9 +85,9 @@ function parseEndpoints(
 
   assertKeyTypes(prefix, ep, {
     region: "array",
-    platform: "string",
+    platform: (platform) => backend.AllFunctionsPlatforms.includes(platform),
     entryPoint: "string",
-    availableMemoryMb: "number",
+    availableMemoryMb: (mem) => backend.AllMemoryOptions.includes(mem),
     maxInstances: "number",
     minInstances: "number",
     concurrency: "number",
@@ -94,7 +95,7 @@ function parseEndpoints(
     timeoutSeconds: "number",
     vpc: "object",
     labels: "object",
-    ingressSettings: "string",
+    ingressSettings: (setting) => backend.AllIngressSettings.includes(setting),
     environmentVariables: "object",
     secretEnvironmentVariables: "array",
     httpsTrigger: "object",
@@ -102,7 +103,16 @@ function parseEndpoints(
     eventTrigger: "object",
     scheduleTrigger: "object",
     taskQueueTrigger: "object",
+    blockingTrigger: "object",
+    cpu: (cpu: backend.Endpoint["cpu"]) => typeof cpu === "number" || cpu === "gcf_gen1",
   });
+  if (ep.vpc) {
+    assertKeyTypes(prefix + ".vpc", ep.vpc, {
+      connector: "string",
+      egressSettings: (setting) => backend.AllVpcEgressSettings.includes(setting),
+    });
+    requireKeys(prefix + ".vpc", ep.vpc, "connector");
+  }
   let triggerCount = 0;
   if (ep.httpsTrigger) {
     triggerCount++;
@@ -117,6 +127,9 @@ function parseEndpoints(
     triggerCount++;
   }
   if (ep.taskQueueTrigger) {
+    triggerCount++;
+  }
+  if (ep.blockingTrigger) {
     triggerCount++;
   }
   if (!triggerCount) {
@@ -192,6 +205,13 @@ function parseEndpoints(
         });
       }
       triggered = { taskQueueTrigger: ep.taskQueueTrigger };
+    } else if (backend.isBlockingTriggered(ep)) {
+      requireKeys(prefix + ".blockingTrigger", ep.blockingTrigger, "eventType");
+      assertKeyTypes(prefix + ".blockingTrigger", ep.blockingTrigger, {
+        eventType: "string",
+        options: "object",
+      });
+      triggered = { blockingTrigger: ep.blockingTrigger };
     } else {
       throw new FirebaseError(
         `Do not recognize trigger type for endpoint ${id}. Try upgrading ` +